Reorganize implementation files
