+++ src/events.cpp
@@ -1,22 +1,10 @@
-#include "ReactESP.h"
+#include "events.h"
 
-#include <Arduino.h>
-#include <FunctionalInterrupt.h>
+#include <freeRTOS/semphr.h>
 
-#include <cstring>
+#include "event_loop.h"
 
 namespace reactesp {
-
-/**
- * @brief Return the current time since the device restart in microseconds
- *
- * Returns the time since the device restart. Even though the time
- * is in microseconds, a 64-bit integer is all but guaranteed not to
- * rewrap, ever.
- */
-#ifdef ESP32
-uint64_t ICACHE_RAM_ATTR micros64() { return esp_timer_get_time(); }
-#endif
 
 // Event classes define the behaviour of each particular
 // Event
@@ -119,91 +107,5 @@ void ISREvent::remove(EventLoop* event_loop) {
   delete this;
   xSemaphoreGiveRecursive(event_loop->isr_event_list_mutex_);
 }
-
-void EventLoop::tickTimed() {
-  xSemaphoreTakeRecursive(timed_queue_mutex_, portMAX_DELAY);
-  const uint64_t now = micros64();
-  TimedEvent* top = nullptr;
-
-  while (true) {
-    if (timed_queue.empty()) {
-      break;
-    }
-    top = timed_queue.top();
-    if (!top->isEnabled()) {
-      timed_queue.pop();
-      delete top;
-      continue;
-    }
-    const uint64_t trigger_t = top->getTriggerTimeMicros();
-    if (now >= trigger_t) {
-      timed_queue.pop();
-      top->tick(this);
-    } else {
-      break;
-    }
-  }
-  xSemaphoreGiveRecursive(timed_queue_mutex_);
-}
-
-void EventLoop::tickUntimed() {
-  xSemaphoreTakeRecursive(untimed_list_mutex_, portMAX_DELAY);
-  for (UntimedEvent* re : this->untimed_list) {
-    re->tick(this);
-  }
-  xSemaphoreGiveRecursive(untimed_list_mutex_);
-}
-
-void EventLoop::tick() {
-  tickUntimed();
-  tickTimed();
-}
-
-DelayEvent* EventLoop::onDelay(uint32_t delay, react_callback callback) {
-  auto* dre = new DelayEvent(delay, callback);
-  dre->add(this);
-  return dre;
-}
-
-DelayEvent* EventLoop::onDelayMicros(uint64_t delay,
-                                       react_callback callback) {
-  auto* dre = new DelayEvent(delay, callback);
-  dre->add(this);
-  return dre;
-}
-
-RepeatEvent* EventLoop::onRepeat(uint32_t interval, react_callback callback) {
-  auto* rre = new RepeatEvent(interval, callback);
-  rre->add(this);
-  return rre;
-}
-
-RepeatEvent* EventLoop::onRepeatMicros(uint64_t interval,
-                                         react_callback callback) {
-  auto* rre = new RepeatEvent(interval, callback);
-  rre->add(this);
-  return rre;
-}
-
-StreamEvent* EventLoop::onAvailable(Stream& stream, react_callback callback) {
-  auto* sre = new StreamEvent(stream, callback);
-  sre->add(this);
-  return sre;
-}
-
-ISREvent* EventLoop::onInterrupt(uint8_t pin_number, int mode,
-                                   react_callback callback) {
-  auto* isrre = new ISREvent(pin_number, mode, callback);
-  isrre->add(this);
-  return isrre;
-}
-
-TickEvent* EventLoop::onTick(react_callback callback) {
-  auto* tre = new TickEvent(callback);
-  tre->add(this);
-  return tre;
-}
-
-void EventLoop::remove(Event* event) { event->remove(this); }
 
 }  // namespace reactesp
